+++ mediorum/server/helpers.go
@@ -1,10 +1,7 @@
 package server
 
 import (
-	"io"
-	"net/http"
 	"net/url"
-	"os"
 	"strconv"
 	"strings"
 
@@ -43,25 +40,6 @@ func (ss *MediorumServer) getScheme() string {
 func cleanHost(host string) string {
 	u, _ := url.Parse(host)
 	return u.Host
-}
-
-func sniffMimeType(r io.ReadSeeker) string {
-	buffer := make([]byte, 512)
-	r.Read(buffer)
-	r.Seek(0, 0)
-	return http.DetectContentType(buffer)
-}
-
-// returns true if file exists and sniffed mime type is audio.*
-// returns false if anything goes wrong (i.e. file doesn't exist).
-// if file doesn't exist we want to fall thru to redirect behavior instead of blocking.
-func isAudioFile(filePath string) bool {
-	if f, err := os.Open(filePath); err == nil {
-		mime := sniffMimeType(f)
-		f.Close()
-		return strings.HasPrefix(mime, "audio")
-	}
-	return false
 }
 
 type ByteRange struct {
