+++ mediorum/server/serve_upload.go
@@ -324,15 +324,6 @@ func (ss *MediorumServer) postUpload(c echo.Context) error {
 	return c.JSON(200, uploads)
 }
 
-// TODO: remove after confirming files didn't start go missing
-func (ss *MediorumServer) getBlobByJobIDAndVariantDeprecated(c echo.Context) error {
-	jobID := c.Param("jobID")
-	variant := c.Param("variant")
-	c.SetParamNames("dirCid", "fileName")
-	c.SetParamValues(jobID, variant)
-	return ss.serveLegacyDirCid(c)
-}
-
 func (ss *MediorumServer) getBlobByJobIDAndVariant(c echo.Context) error {
 	// if the client provided a filename, set it in the header to be auto-populated in download prompt
 	filenameForDownload := c.QueryParam("filename")
